requirements-dev.txt
matplotlib
numpydoc
oct2py
pytest
sphinx
